+++ components/core/ui/timeline/src/commonMain/kotlin/com/flipperdevices/ui/timeline/TimelineComposablePreview.kt
@@ -1,0 +1,57 @@
+package com.flipperdevices.ui.timeline
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.Surface
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.platform.LocalDensity
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.flipperdevices.bsb.core.theme.BusyBarThemeInternal
+import kotlin.time.Duration.Companion.hours
+import kotlin.time.Duration.Companion.seconds
+import org.jetbrains.compose.ui.tooling.preview.Preview
+
+
+@Preview
+@Composable
+fun TimelineComposablePreview() {
+    BusyBarThemeInternal {
+        val numSegments = 5
+        val state = rememberPodcastSliderState(
+            currentValue = 75.seconds.inWholeSeconds.toFloat(),
+            range = 0..9.hours.inWholeSeconds.toInt()
+        )
+        val animatedTextState = remember { AnimatedTextState(24f, 0.4f) }
+        Surface(modifier = Modifier.fillMaxWidth(), color = Color.Black) {
+            PodcastSlider(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(vertical = 16.dp),
+                density = with(LocalDensity.current) { 25.dp.toPx().toInt() },
+                numSegments = numSegments,
+                state = state,
+                indicatorLabel = { alpha, value ->
+                    if (value % 5 == 0) {
+                        Box(
+                            modifier = Modifier.height(with(LocalDensity.current) { 85.sp.toDp() }),
+                            contentAlignment = Alignment.Center
+                        ) {
+                            TextContent(
+                                value = value,
+                                isSelected = state.selected.value == value,
+                                animatedTextState = animatedTextState
+                            )
+                        }
+                    }
+                }
+            )
+        }
+    }
+}
